% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fruit_and_veggies.R
\docType{data}
\name{fruit_and_veggies}
\alias{fruit_and_veggies}
\alias{fruit_lbl}
\alias{fruit_lbl_int}
\alias{fruit_lbl_chr}
\alias{fruit_fct}
\alias{veggies}
\alias{exotic_veggies}
\title{Example Data Sets: Fruits and Vegetables}
\format{
\code{fruit_lbl} is a labelled vector of double type. It includes
different fruit names as labels.

\code{fruit_lbl_int} is similar to \code{fruit_lbl}, but it's of integer type.
It illustrates how labels behave with integer data.

\code{fruit_lbl_chr} is another variant of \code{fruit_lbl} where labels are
assigned to characters instead of numbers, illustrating how string labels
work.

\code{fruit_fct} is a factor variant of the fruit dataset. It helps
demonstrate the difference between factors and labelled vectors.

\code{veggies} is a labelled vector of double type with labels
assigned to different vegetables, serving as a basic vegetable dataset.
Rather than running from \code{1:n}, the values are prime numbers, illustrating
how labelled variables are more flexible than factors regarding the
underlying values behind the labels. This variable also has a \code{var_label}.

\code{exotic_veggies} is an extension of \code{veggies} with more exotic
vegetable names as labels, and has a \code{var_label}. More importantly, this
demonstrates how different types of missing values behave with labelled
values. Specifically:
\itemize{
\item There are no observations of \verb{[7] Eggplant} or \verb{[41] Zuchini}.
\item There are five instances of the unlabelled value \verb{[43]} (no corresponding label for this value).
\item There are six instances of \verb{[NA]}, which also does not have a corresponding label.
}
}
\source{
These variables were generated as illustrative example data.
}
\usage{
fruit_lbl

fruit_lbl_int

fruit_lbl_chr

fruit_fct

veggies

exotic_veggies
}
\description{
A collection of labelled vectors and factors representing
different fruits and vegetables. See the format section below for specifics
about the class and format of each variable.
}
\seealso{
\link[haven]{labelled}
}
\keyword{datasets}
