% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{lbl_assert_filled}
\alias{lbl_assert_filled}
\alias{lbl_assert_dense}
\title{Assert that a labelled vector is filled/dense}
\usage{
lbl_assert_filled(
  x,
  na_unfilled_ok = TRUE,
  .var.name = checkmate::vname(x),
  add = NULL
)

lbl_assert_dense(x, .var.name = checkmate::vname(x), add = NULL)
}
\arguments{
\item{x}{The value to assert.}

\item{na_unfilled_ok}{Should unlabelled \code{NA} values throw an error?}
}
\value{
x unless there is an error
}
\description{
\code{lbl_assert_filled()} asserts that a given labelled vector is \code{filled}, i.e.,
that for every value that is present in the vector data, the corresponding
value is present in the value labels. Not having labels for the \code{NA} value
is OK unless explicitly disallowed.

\code{lbl_assert_dense()} asserts that a given labelled vector is \code{dense}, i.e.,
that for every value that is present in the vector data, the corresponding
value is present in the value labels. \code{NA} values are ignored, as density is not applicable for them.
}
\keyword{internal}
