% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densify.R
\name{lbl_densify}
\alias{lbl_densify}
\title{Densify a labeled vector}
\usage{
lbl_densify(x, unlabelled = c("error", "fill"))
}
\arguments{
\item{x}{A labelled vector to be densified.}

\item{unlabelled}{This parameter determines the handling of unlabelled values
in the input vector. It can take two values:
\itemize{
\item "error": The function will stop and an error message will be returned if
there are unlabelled values in the vector.
\item "fill": The function will fill any unlabelled values using the \code{\link[=lbl_fill]{lbl_fill()}}
function before densifying.
}}
}
\value{
A labeled vector of the same length as the original vector, \code{x}, but with
values replaced with their dense ranks. The labels are correctly associated with
their respective values and the variable label (if present) is preserved.
}
\description{
The \code{lbl_densify} function takes a labeled vector and assigns new,
densely-ranked integer values to each unique value in the vector. This is
particularly useful when working with sparse or unevenly distributed data, as
it transforms the data into a format where each unique value is represented
by a unique integer, starting from 1 and increasing by 1 for each new unique
value. The order of the unique values is preserved. This function keeps the
labels correctly associated with their respective values throughout the
densifying process.
}
\examples{
  # Densify the veggies vector, then examine the result with lbl_count()
  lbl_densify(veggies)
  lbl_densify(veggies) |> lbl_count()

}
