% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop.R
\name{lbl_drop}
\alias{lbl_drop}
\title{Drop unused labels in a labelled vector}
\usage{
lbl_drop(x)
}
\arguments{
\item{x}{A labeled vector}
}
\value{
A labeled vector of the same length as x, without unused labels.
}
\description{
The \code{lbl_drop} function takes a labelled vector and removes labels that are
not associated with any value in the vector.
}
\examples{
  # Drop unused labels in exotic_veggies
  lbl_drop(exotic_veggies)

  # Fewer observations, fewer used labels
  lbl_drop(exotic_veggies[1:10])

}
\seealso{
\code{\link[forcats:fct_drop]{forcats::fct_drop()}}
}
