% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill.R
\name{lbl_fill}
\alias{lbl_fill}
\title{Fill unlabeled values in a labeled vector}
\usage{
lbl_fill(x, fill_na = FALSE, pattern = "<{v}>")
}
\arguments{
\item{x}{A labeled vector}

\item{fill_na}{Should NA values be filled}

\item{pattern}{A glue pattern determining the text for the new labels}
}
\value{
A labeled vector of the same length as x, with labels for all values.
}
\description{
The \code{lbl_fill} function takes a labeled vector and creates new labels for any
values that are present in the vector but do not have an existing label.
}
\examples{
# Fill the unlabelled values in the exotic_veggies example variable
lbl_fill(exotic_veggies)

}
