% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift.R
\name{lbl_shift}
\alias{lbl_shift}
\title{Shift Levels of a Labelled Vector}
\usage{
lbl_shift(x, n = -1L, rotate = FALSE)
}
\arguments{
\item{x}{A labelled vector.}

\item{n}{An integer. Positive values shift up (increasing the values);
negative values shift to town (decreasing the value). Default is -1L, which
would shift one-based label set to a zero-based label set.}

\item{rotate}{A logical. If TRUE, it would perform a rotating shift on the
labels. However, this feature is currently not implemented. Default is
FALSE.}
}
\value{
A \code{labelled} object with shifted labels.
}
\description{
This function shifts the levels of a labelled vector to the left or right.
Unlike \code{forcats::fct_shift()} from the \code{forcats} package, which works with
factors, this function is designed to work with haven's labelled vectors.
Given that labelled vectors can have any values, shifting simply increases or
decreases the values of the vector, updating the labels to match. Shifting by
a positive number thus shifts the value \emph{up} rather than down, as opposed to
how \code{fct_shift()} works. However, maintaining some compatibility, the default
value of n is \code{-1L}, meaning that one-based vector gets shifted to a
zero-based vector, which is often useful.

Implementing a rotating shift (as \code{fct_shift()} does) would be possible, but
has not been done yet. For \verb{labelled<character>}, only rotating shifts make
sense, and so shifting such vectors is not currently supported.
}
\examples{
  lbl_shift(fruit_lbl)
  lbl_shift(fruit_lbl_int, 1)
  lbl_shift(veggies, 20)

}
\seealso{
\code{\link[forcats:fct_shift]{forcats::fct_shift()}}
}
