% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_labels.R
\name{lbl_sort_labels}
\alias{lbl_sort_labels}
\title{Sort value labels of a labelled variable}
\usage{
lbl_sort_labels(x)
}
\arguments{
\item{x}{A \code{labelled} variable.}
}
\value{
A \code{labelled} variable with values and label levels equal to \code{x}, but
with inner representation of the value labels in ascending order.
}
\description{
Sorts the labels of a \code{labelled} variable so that the inner representation of
the vector is ascending.
}
\examples{
lbl_sort_labels(fruit_lbl)
lbl_sort_labels(veggies)
lbl_sort_labels(exotic_veggies)

}
