% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{as_labelled}
\alias{as_factor}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{forcats}{\code{\link[forcats]{as_factor}}}

  \item{sjlabelled}{\code{\link[sjlabelled]{as_labelled}}}
}}

