% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/001-forlabs-package.R
\docType{package}
\name{forlabs-package}
\alias{forlabs}
\alias{forlabs-package}
\title{forlabs: Tools for Working with Labelled Variables}
\description{
Provides functions for labelled variables (`haven_labelled`) that correspond to those that the `forcats` package provides for categorical variables (`factor`).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/torfason/forlabs}
  \item \url{https://torfason.github.io/forlabs/}
  \item Report bugs at \url{https://github.com/torfason/forlabs/issues}
}

}
\author{
\strong{Maintainer}: Magnus Thor Torfason \email{m@zulutime.net}

}
\keyword{internal}
