% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{lbl_count}
\alias{lbl_count}
\title{Count entries in a labelled variable}
\usage{
lbl_count(x, sort = FALSE, prop = FALSE, ..., include_var_label = FALSE)
}
\arguments{
\item{x}{A labelled variable}

\item{sort}{If \code{TRUE}, the result will be sorted so that the most common
values float to the top. If it is false, the result will be sorted in the
order of the \emph{values} of the labelled variable.}

\item{prop}{If \code{TRUE}, compute the fraction of marginal table.}

\item{...}{Currently unused}

\item{include_var_label}{Should an extra line be included with the variable
label (if any).}
}
\value{
A \code{tibble} with columns \code{v}, \code{l}, \code{n} and \code{p}, if prop is \code{TRUE}.
}
\description{
Count entries in a labelled variable
}
\examples{
  table(fruit_lbl)
  lbl_count(fruit_lbl)
  lbl_count(fruit_lbl, sort = TRUE)
  lbl_count(fruit_lbl, sort = TRUE, prop = TRUE)

}
\seealso{
\code{\link[forcats:fct_count]{forcats::fct_count()}}
}
