% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop.R
\name{lbl_drop}
\alias{lbl_drop}
\title{Drop unused labels in a labelled vector}
\usage{
lbl_drop(x)
}
\arguments{
\item{x}{A labeled vector}
}
\value{
A labeled vector of the same length as x, without unused labels.
}
\description{
The \code{lbl_drop} function takes a labelled vector and removes labels that are
not associated with any value in the vector.
}
\examples{
# Create labeled vector with unused labels
lbl_drop(exotic_veggies)

}
