% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{lbl_recode}
\alias{lbl_recode}
\title{Recode labelled variables using tilde notation}
\usage{
lbl_recode(.x, ..., .sep = " / ")
}
\arguments{
\item{.x}{The \code{labelled} vector to recode.}

\item{...}{A recode specification in \code{tilde} (\code{formula}) syntax.}

\item{.sep}{A string for how to separate labels that need to be combined.}
}
\value{
The re-coded vector.
}
\description{
Recodes labelled variables and takes care to recode the levels of the value
labels so that the labels continue to track the underlying values. Recoding
multiple values to the same value results in corresponding variable labels
being combined. Recoding to \code{NA} results in corresponding labels to be
dropped since \code{NA} should not be labelled in standard \code{haven_labelled}
variables.

The underlying recoding takes place using \code{dplyr::case_match()}, which takes
\code{tilde} or \code{formula} notation.
}
\examples{
lbl_recode(fruit_lbl, 1:2 ~ 1, 3 ~ 2, 4:5 ~ 3)
lbl_recode(fruit_lbl_chr, "d" ~ "z", "e" ~ NA)
lbl_recode(exotic_veggies, 3:10 ~ 1, 11:20 ~ 2, 21:41 ~ 3,
                           2 ~ NA, NA ~ 99)

}
