% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/001-forlabs-package.R,
%   R/003-assertions-checkmate-imports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{labelled}
\alias{vec_data}
\alias{vec_c}
\alias{ll_assert_labelled}
\alias{ll_to_character}
\alias{ll_labelled}
\alias{ll_val_labels}
\alias{ll_var_label}
\alias{qassert}
\alias{assert_flag}
\alias{assert_int}
\alias{assert_number}
\alias{assert_string}
\alias{assert_scalar}
\alias{assert_count}
\alias{assert_true}
\alias{assert_false}
\alias{assert_null}
\alias{assert_scalar_na}
\alias{assert_atomic}
\alias{assert_atomic_vector}
\alias{assert_character}
\alias{assert_complex}
\alias{assert_date}
\alias{assert_double}
\alias{assert_factor}
\alias{assert_integer}
\alias{assert_integerish}
\alias{assert_logical}
\alias{assert_numeric}
\alias{assert_posixct}
\alias{assert_raw}
\alias{assert_vector}
\alias{assert_list}
\alias{assert_named}
\alias{assert_names}
\alias{assert_array}
\alias{assert_matrix}
\alias{assert_data_frame}
\alias{assert_data_table}
\alias{assert_tibble}
\alias{assert_class}
\alias{assert_environment}
\alias{assert_formula}
\alias{assert_function}
\alias{assert_multi_class}
\alias{assert_r6}
\alias{assert_choice}
\alias{assert_disjunct}
\alias{assert_permutation}
\alias{assert_set_equal}
\alias{assert_subset}
\alias{assert_access}
\alias{assert_directory}
\alias{assert_file}
\alias{assert_os}
\alias{assert_path_for_output}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{checkmate}{\code{\link[checkmate:checkAccess]{assert_access}}, \code{\link[checkmate:checkArray]{assert_array}}, \code{\link[checkmate:checkAtomic]{assert_atomic}}, \code{\link[checkmate:checkAtomicVector]{assert_atomic_vector}}, \code{\link[checkmate:checkCharacter]{assert_character}}, \code{\link[checkmate:checkChoice]{assert_choice}}, \code{\link[checkmate:checkClass]{assert_class}}, \code{\link[checkmate:checkComplex]{assert_complex}}, \code{\link[checkmate:checkCount]{assert_count}}, \code{\link[checkmate:checkDataFrame]{assert_data_frame}}, \code{\link[checkmate:checkDataTable]{assert_data_table}}, \code{\link[checkmate:checkDate]{assert_date}}, \code{\link[checkmate:checkDirectoryExists]{assert_directory}}, \code{\link[checkmate:checkDisjunct]{assert_disjunct}}, \code{\link[checkmate:checkDouble]{assert_double}}, \code{\link[checkmate:checkEnvironment]{assert_environment}}, \code{\link[checkmate:checkFactor]{assert_factor}}, \code{\link[checkmate:checkFALSE]{assert_false}}, \code{\link[checkmate:checkFileExists]{assert_file}}, \code{\link[checkmate:checkFlag]{assert_flag}}, \code{\link[checkmate:checkFormula]{assert_formula}}, \code{\link[checkmate:checkFunction]{assert_function}}, \code{\link[checkmate:checkInt]{assert_int}}, \code{\link[checkmate:checkInteger]{assert_integer}}, \code{\link[checkmate:checkIntegerish]{assert_integerish}}, \code{\link[checkmate:checkList]{assert_list}}, \code{\link[checkmate:checkLogical]{assert_logical}}, \code{\link[checkmate:checkMatrix]{assert_matrix}}, \code{\link[checkmate:checkMultiClass]{assert_multi_class}}, \code{\link[checkmate:checkNamed]{assert_named}}, \code{\link[checkmate:checkNames]{assert_names}}, \code{\link[checkmate:checkNull]{assert_null}}, \code{\link[checkmate:checkNumber]{assert_number}}, \code{\link[checkmate:checkNumeric]{assert_numeric}}, \code{\link[checkmate:checkOS]{assert_os}}, \code{\link[checkmate:checkPathForOutput]{assert_path_for_output}}, \code{\link[checkmate:checkPermutation]{assert_permutation}}, \code{\link[checkmate:checkPOSIXct]{assert_posixct}}, \code{\link[checkmate:checkR6]{assert_r6}}, \code{\link[checkmate:checkRaw]{assert_raw}}, \code{\link[checkmate:checkScalar]{assert_scalar}}, \code{\link[checkmate:checkScalarNA]{assert_scalar_na}}, \code{\link[checkmate:checkSetEqual]{assert_set_equal}}, \code{\link[checkmate:checkString]{assert_string}}, \code{\link[checkmate:checkSubset]{assert_subset}}, \code{\link[checkmate:checkTibble]{assert_tibble}}, \code{\link[checkmate:checkTRUE]{assert_true}}, \code{\link[checkmate:checkVector]{assert_vector}}, \code{\link[checkmate]{qassert}}}

  \item{haven}{\code{\link[haven]{labelled}}}

  \item{vctrs}{\code{\link[vctrs]{vec_c}}, \code{\link[vctrs]{vec_data}}}

  \item{zmisc}{\code{\link[zmisc]{ll_assert_labelled}}, \code{\link[zmisc]{ll_labelled}}, \code{\link[zmisc]{ll_to_character}}, \code{\link[zmisc]{ll_val_labels}}, \code{\link[zmisc]{ll_var_label}}}
}}

